Fixed stuff related to the Slider button
Fixed the sliders being outside of the Menu and started working on keeping the text in the middle of the slider (WIP, needs refactoring and a new approach, lol).
Also applied the new slider button to the gravity on the World Options Menu.
Added a new color on the slider button to match the "infamous" theme.