Minor slider (fix?) and started a ped costumization menu
Sliders still break a bit when their index is bigger than 50-70, yet they seem to be equally broken as for an index of 50 and an index of 200. This hotfix should make it less obvious but that code needs to be rewritten with bigger values in mind.

Started a Ped Costumization Menu where you will be able to change everything about your ped.